refctoring card Code
